requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# Authentication & Security
PyJWT==2.8.0
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.4
Pillow==10.0.0

# Payments
stripe==7.4.0
requests==2.31.0

# Utilities
qrcode==7.4.2
marshmallow==3.20.1
alembic==1.13.2
blinker==1.7.0
